# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Code Red Backend API
  version: 1.3.0
  description: >
    Production-style backend contract for the Code Red coaching institute platform.
    Canonical routes are versioned under `/api/v1`. This revision aligns the contract
    with the current frontend implementation, including authority-based navigation,
    inventory requests, payroll views, promotion flows, and faculty material publishing.
    Access control follows the frontend role families: student, faculty, parent,
    and administration (`admin`, `administration`, `director`, `superadmin`). Request
    payloads are validated with Pydantic, some background work is dispatched through
    Celery using Redis, and throttled endpoints may return HTTP 429.
servers:
  - url: http://localhost:3500
    description: Local Flask API server
  - url: http://localhost:3000
    description: Frontend dev server proxy, if configured
...
paths:
...
  /api/v1/classes:
    get:
      tags: [Academics]
      summary: List classes
      description: Requires a bearer token entered through the Authorize button. This endpoint does not accept query or path parameters.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Class list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClassInfo'
...
  /api/v1/faculty:
    get:
      tags: [Faculty]
      summary: List faculty
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Faculty list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Faculty'
  /api/v1/faculty/classes:
    get:
      tags: [Faculty]
      summary: List faculty classes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Faculty class list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FacultyClass'
  /api/v1/faculty/classes/{classId}/subjects:
    get:
      tags: [Faculty]
      summary: List subjects for a faculty-visible class
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ClassIdPath'
      responses:
        '200':
          description: Subject list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subject'
  /api/v1/faculty/subjects/{subjectId}/materials:
    get:
      tags: [Faculty]
      summary: List materials for a subject
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SubjectIdPath'
      responses:
        '200':
          description: Materials list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Material'
        '404':
          $ref: '#/components/responses/Error404'
    post:
      tags: [Faculty]
      summary: Publish study material for a subject
      description: >
        Creates a new study material record for the selected subject. Materials may be
        grouped week-wise or chapter-wise and should become visible in the student subject page.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SubjectIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialCreateRequest'
      responses:
        '201':
          description: Material published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Material'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  parameters:
...
    ClassIdPath:
      in: path
      name: classId
      required: true
      schema: { type: string }
    SubjectIdPath:
      in: path
      name: subjectId
      required: true
      schema: { type: string }
...
  responses:
...
    Error403:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Error404:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: object
              additionalProperties: true
...
    Faculty:
      type: object
      properties:
        id: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string, format: email }
        subjects:
          type: array
          items: { type: string }
    FacultyClass:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        section: { type: string, nullable: true }
    Subject:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        code: { type: string, nullable: true }
        teacherId: { type: string, nullable: true }
        facultyId: { type: string, nullable: true }
        credits: { type: integer, nullable: true }
    ClassInfo:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        level: { type: integer }
        section: { type: string, nullable: true }
...
    Material:
      type: object
      properties:
        id: { type: string }
        subjectId: { type: string, nullable: true }
        subjectName: { type: string, nullable: true }
        chapterId: { type: string, nullable: true }
        chapterTitle: { type: string, nullable: true }
        title: { type: string }
        unit: { type: string, nullable: true }
        week: { type: string, nullable: true }
        type: { type: string }
        description: { type: string, nullable: true }
        fileName: { type: string, nullable: true }
        documentId: { type: string, nullable: true }
        documentName: { type: string, nullable: true }
        documentUrl: { type: string, nullable: true }
        externalUrl: { type: string, nullable: true }
        uploadedAt: { type: string, format: date-time, nullable: true }
        storagePath: { type: string, nullable: true }
        uploadedBy: { type: string, nullable: true }
        isFacultyUpload: { type: boolean, nullable: true }
...
    MaterialCreateRequest:
      type: object
      required: [title, type]
      properties:
        title: { type: string }
        type: { type: string }
        description: { type: string, nullable: true }
        week: { type: string, nullable: true }
        unit: { type: string, nullable: true }
        chapterId: { type: string, nullable: true }
        chapterTitle: { type: string, nullable: true }
        fileName: { type: string, nullable: true }